Extract PresenceMap to its own file for unit-testability
